requirements.txt
faiss-cpu
PyPDF2
python-docx
requests
google-generativeai
python-dotenv
fastapi
uvicorn 